docs: link to the api-ref for 3.27/3.44 for the version history
The REST API version history should ideally have some detailed
explanation of the changes introduced in that microversion. Barring
that, at least link to the API reference (if it exists) so client
side users have a breadcrumb trail to figure out what that microversion
introduces. That's what I'm doing here.

Change-Id: If15023b4ff9254fa6a7efa37dd132d883570faa0
Closes-Bug: #1716705