Check for outstanding attachments during reserve
There are a number of places on the Nova side where we need to do things
with volume connections like remove an attachment and immediately
reserve a new one. This introduces a potential race where bad things
can happen. To deal with that, we actually want to to create a new
attachment to do a reserve, BEFORE we delete the current attachment.

The problem is that the new Attachment code currently will toggle the
volume status and attach status such that you will then not be able to
remove the attachment in the future, and it also makes things such that
the true state of the volume is not reflected properly.

This patch adds a check on reserve for additional attachments and if
there are attachments for a volume that have the volume as in-use we
don't clear that status during the new reserve call.

Change-Id: I8062897bd3d4fe20de9a5660eac6fec856cc3c1e
Closes-Bug: #1717564