[DOC] Update index page for Install tutorial
The index page for our install tutorial should point
back to the top level OpenStack installation tutorial
to give the user information on how to set up the base
OpenStack environment before starting installation.

This patch adds the link and notes that the setup in
there is a prerequisite to executing the Cinder Install
Tutorial.

Change-Id: I42418b30b3978bac24a856c22342ce5db3e42282