VMAX driver - Volume identifier issues
When creating a VMAX volume through cinder, a 'volume_identifier' is
set on the device on the backend. This takes the form 'OS-<cinderUUID>'.
There are a couple of issues with how this is currently being done:
1. When the VMAX driver retrieves the volume using the 'device_id' value
saved in the provider location, it checks that the volume identifier
matches OS-<cinderUUID>. However, the way this is currently being
checked can cause an issue if the create volume operation ends up being
retried in cinder - sometimes there can be two devices on the backend
with the same volume identifier, and the code may not pick up the
correct device.
2. On a deallocate operation, the volume identifier is not being unset
on the device. This patch rectifies these issues.

Change-Id: I8218768216e87f6fdda1dd77dd09ca2ca46111c7
Closes-Bug: 1717933