VMAX driver - Remove workload for next gen arrays
The next release of VMAX no longer has an option to set 'Workload' on a
storage group. Currently, the VMAX driver expects a workload to be set
when the user creates a VMAX volume type, which will cause an issue for
users using the next gen VMAX arrays. This patch rectifies the issue.

Change-Id: I8ad7fbdca5a2695734b118d60a29a71064ae9b0b
Closes-Bug: 1717289