Stop overriding LVM overprovisioning ratio and deprecate
The LVM driver has its own config option for managing overprovisioning
separate from our common overprovisioning ratio. It default to 1,
causing it to always override the common default.

This changes the LVM default to None so it does not override the common
setting and marks this config option as deprecated to be removed in
Rocky.

Change-Id: Ib90cdb1f60741284ea171d1a84ad74beac43c7d3
Closes-bug: #1720004