Kaminario K2: Support duplicated FQDN in network
The Kaminario K2 driver uses the FQDN of the node that is doing the
attach as an unique identifier to map the volume.

The problem is that the FQDN is not always unique, there are
environments where the same FQDN can be found in different systems, and
in those cases if both try to attach volumes the second system will
fail.

One example of this happening would be on a QA environment where you are
creating VMs and they all have names like controller-0.localdomain and
compute-0.localdomain.

This patch adds a configuration option to the K2 driver called
`unique_fqdn_network` to support these kind of environments.

Closes-Bug: #1720147
Change-Id: I2099b804a2043078e50a5a9e14c1497861576f61