Add uuid to services entries
Our services ID column is still just an integer and that's *ok* for the
most part.  There are things that it would be nice to advertise a
more meaningful and unique identifier to the users of services though.

This change adds an indexable UUID column to the services, this will
be useful for things like identifying backends without leaking
details to end users.

Change-Id: I67e52c6a8634b74bf5975290298d6fbcadc7dd50