NetApp E-series: Fix provisioned_capacity_gb
Currently, NetApp E-series drivers (iSCSI and FC) are incorrectly
reporting the total physical used size as the provisioned_capacity_gb.
This patch fixes the calculation of this attribute by summing the
user-visible capacity of all the volumes present in the backend and
reporting the result as the provisioning_capacity_gb.

Change-Id: I1471ecfffe340dafb1c95bac5eaff79cde650955
Closes-Bug: #1718739