Fix Tempest Volume Revert test
When we stopped using magic strings for the microversions in Cinder on
change Ib3a80fee6caaabb49af097aa197f550c65d94985 we broke the Volume
Revert Tempest test because we started sending requests to do the revert
with API microversion 3.0 and the minimum version is 3.40.

This patch fixes it.

Closes-bug: 1720855
Change-Id: Ib85411c7f6f5cc9ad33e41f1aba18021703e4913