NetApp ONTAP: Remove support for 7mode systems
The Unified Driver for NetApp Storage in Cinder
supports two families of ONTAP, 7mode and Clustered
Data ONTAP.

ONTAP 7 is now officially nearing the end-of-life
of product support. The deprecation notice [1]
for these drivers in Cinder was issued in the Newton
Release, so it is time to remove them from tree.

[1] http://lists.openstack.org/pipermail/openstack-operators/2016-November/011957.html

Implements: bp remove-netapp-7mode-drivers

Change-Id: I129ca060a89275ffd56481b8f64367b0d803cff5