Update BlockBox
There have been a number of changes to LOCI since the introduction
of BlockBox.  In particular LOCI now offers a `universal` Dockerfile
and employees bindeps and new ARGS features of Docker to customize
images.

This patch modifies BlockBox to use the new universal LOCI repo and
also cleans up a bunch of cruft and adds an updated conf entry with
key-manager.

This patch also removes the dev environment as it's out dated and
doesn't seem to be gaining much usage.  It's easy enough to set
this up on your own, so rather than maintain it just remove it.

Change-Id: Id03ff9310fd53c4e6df300f2866b7a17cba82561