VMAX driver - detach volume shouldn't remove from volume groups
Detaching a volume in VMAX removes the volume from all its current
storage groups and returns it to its default storage group. This is not
desired behaviour when the volume is a member of a generic volume group.
The volume should only be removed from the storage group associated
with the masking view related to the detach operation (or all masking
views in the case of a force detach). This patch rectifies the issue.

Change-Id: I065ffba9615af54998ae94a8d2d2fd3853f462cb
Closes-Bug: 1721207
(cherry picked from commit 27fd333df9a7f5b473225b3c55cd0c250f017134)