SMBFS: manageable volumes
This change allows the SMBFS driver to claim (manage) vhd/x images
present on pre-configured shares.

Note that images having backing (parent) files will be rejected.
Managing snapshots will not be supported either.

The images may reside in subdirectories (and get moved afterwards),
which may be useful when having a folder dedicated to images that
are to be imported in Cinder.

All the logic is included in the new "RemoteFSManageableVolumesMixin"
class so that any related driver may inherit it. For the record,
this works out of the box for the NFS driver (special care has been
taken in this sense).

Implements: blueprint remotefs-manage-unmanage-volume
Change-Id: I54695655e563d84e4fb1b76c42f0127c5fb909f7