DS8000 ConsistencyGroup failed to update group with volumes
While adding a volume to ConsistencyGroup the failed with TypeError: sequence
item 0: expected string, tuple found.  This could have happened if the driver
did not find the specified LSS pairs matching that of the available volume
pools.  The driver intended to throw an error message with available_lss_pairs
but failed to concatinate the data to the error message with TypeError
exception.

The proposed fix attempts to address this issue by simplifying the
concat of available_lss_pairs as string and then adding to error message.

Closes-bug: #1722769

Change-Id: I4c154b15991fd624d91bd3e3472d25c714fb8525
(cherry picked from commit 75a1849c9854cd5f8d366de6087e50eff9b595e4)