Deprecate SolidFire Image Caching feature
We've had a general solution to image-caching in
Cinder for a while now and it works well, so we
can remove the SolidFire specific implementation
in favor of using the general solution.

https://docs.openstack.org/cinder/latest/\
admin/blockstorage-image-volume-cache.html

This patch marks the SolidFire config option
as deprecated, and we'll remove the feature
altogether in the next release.

Change-Id: Iefc207955e4718a7d9c3b235b133a4c629977d03