Add display_name to solidfire volume attributes
This just adds another mapping to the SolidFire volume
attributes during create.  We add a `cinder-name` key
to the internal SolidFire volume attributes which has a
value set to the cinder display_name attribute of the volume.

In the case of a volume with no display_name we set the
attribute value to an empty string.

Change-Id: Idbfbf8be7ef34c134c7723891aba8b3baa2ab43a