Enable using extra-specs for SF attributes
There a number of use cases where it's desirable to do things with volumes
out of band (ie not under Cinders control). One example is external backups
without the use of c-backup.

It's pretty simple to signify these sorts of things from Cinder through
the use of extra-specs keys. For example, create a type 'daily-backup'
with extra-specs 'SFAttribute:backup=daily' which may instruct some automation
to query the SolidFire device for any volumes with the attribute
"backup=daily" to be backed up on a daily schedule.

These attributes are just metadata and can be used for anything.
They don't effect Cinder's operation or change anything in how the volume behaves
from Cinders perspective, but instead add additional information and signals for
admins and automation outside of OpenStack/Cinder.

The only requirement is the use of a key delimiter, in this case `SFAttribute`
which instructs the scheduler to ignore the associated key for scheduling, and also
is used as a sentinal for the SF driver to determine what to store in attributes.

Change-Id: Ia02edfb6f30edfccb2bcdd41e8a3fdd7331cef5b
Implements: blueprint enable-extraspec-keys-as-solidfire-attributes