Clean up driver configuration reference
Over the last few cycles we have removed drivers, but we did not always
clean up the documentation from the openstack-manuals project. Those docs
were moved over wholesale as part of the docs migration, so we ended up
with them in the repo with no corresponding driver files.

This removes those leftover docs and cleans up the generated documentation.

Change-Id: Ib9fc4e14333548b84666e3f571e7b42fe20e0319