VMware: Improve scalability of querying volumes
Currently we query volumes in vCenter by name which is not scalable.
Commit 6b9464e91840f6e006f537dcd3c0f656ae90a179 sets the backend
volume ID to Cinder volume ID so that we can use vCenter SearchIndex
API to query backend volumes by Cinder volume ID.

This patch uses SearchIndex API to improve scalability. But this type
of querying will not work for legacy volumes whose vCenter ID is not
set to Cinder volume ID. To handle those volumes, we build a cache to
map their names to vCenter references during driver setup.

Closes-bug: #1600754
Change-Id: I6b49c2590d0bd3946fa03dbac08ab4f778007111