Don't fail when deleting missing backup
The problem here is sometimes when the path doesn't exist and we
try to delete it then exception occurs & breaks the flow. In this
case it happens with the volume delete flow, wherein the exception
occurs just before the quota commit & commit doesn't happen even
though the backup volume got deleted, hence catching the exception
just where it occurs.

Change-Id: I8713f846b1e2e30b4bbc4af57645fdc04bcd8386
Closes-Bug: 1726715