VMAX driver - Incorrect stats reporting
There is an issue in the way free_capacity_gb is being calculated for
the VMAX driver - it is using 'total_allocated_capacity' instead of
calculating using 'total_used_cap', or taking 'fba_free_capacity'. There
is also excess queries being made (and logs being produced), as the
capacities are being requested for each pool. This is unnecessary since
WLP support was removed - the capacities are reported at an SRP level
and so will be the same for each service level/ workload combination.
This patch rectifies these issues.

Change-Id: Icdafd47865e3cf1e5b687f39f15b8ebfccbf67a3
Closes-Bug: 1731930