Fix VolumeAttachment OVO Volume lazy loading
Current code for VolumeAttachment OVO cannot lazy load the volume OVO
correctly and will always return a VolumeNotFound exception.

The reason for not being able to find the volume is that we are using
the VolumeAttachment id to search for the volume instead of the volume's
id.

This patch fixes this by using field "volume_id" instead of field "id"
when calling Volume's get_by_id method.

TrivialFix

Change-Id: Ibe02ca83a674c038ccec28cb104cc2af896426f9