Handle assertRaisesRegexp Python 3 deprecation
In Python 3, assertRaisesRegexp has been deprecated in favor of
assertRaisesRegex. This newer name is not available in Python 2
though, so we can't just change over to the new name.

To get rid of the deprecation messages in our py35 unit test runs,
this adds an alias to our test base that just provides an alias
rather than updating all locations in our tests that use this.

Change-Id: I3cf8e29184a68a07d35fb121c816e40b89df393e