Change default test log level
Despite comments stating otherwise, we set our root logging level
to DEBUG. This causes some unwanted things to be logged to the
console and large test artifacts.

There is also some suspicion that this is causing test failures
with some kind of timing change with stestr causing subunit
parsing failures.

This changes the default level to be higher and fixes a few test
cases that had expected the previous default.

Change-Id: Ib7e45915898347549c12bc8276df556b272bb259
Closes-bug: #1728640