Prevent leaking encryption key when deleting volume
Fail requests to delete an encrypted volume when the associated
encryption key cannot be deleted. This avoids leaking an encryption
key when the volume is deleted but the key cannot be deleted.

This patch addresses situations where Barbican is the key manager,
and Barbican's security policy does not allow the user to delete
the key. Barbican has stricter rules on deleting secrets than Cinder
does on deleting volumes. For example, given two non-admin users X
and Y, Cinder allows user X to delete user Y's volume, but Barbican
does not allow user X to delete user Y's secrets.

If the encryption key cannot be deleted then the volume's status is
set to "error_deleting". To delete the volume, the volume's status
can be reset, and a user with suitable privileges (i.e. an admin or
the volume's owner) can reissue the delete request.

Closes-Bug: #1731528
Change-Id: If0462c18345a4bfbfbcecb6b8caa33d38b86d7f0