Remove v1 API reference
Commit 3e91de956e1947a7014709010b99df380242ac74 removed the
v1 API code that these documents were built from, so they
no longer work. The references to the docs from the main
index just redirect back to the main index, so they can't be
used since the code is gone. Anyone that wants to build the
v1 API reference has to build it locally from before commit
3e91de956e1947a7014709010b99df380242ac74, but even then the
v1 API reference was sorely lacking details on the majority
of the API extensions anyway.

Change-Id: Ie4b25627d962aa7daa684de0d22c6bd212b83e46