VNX: Fix cloning failure when system is busy
In newton, the 30s wait was removed due to refactor, in this case,
If the session hasn't started in the VNX, the driver would perform
unexpected rollback for the migration.

Adding a wait will make sure the session is started on the VNX thus
cloning can succeed.

Change-Id: I3419e628d42eddfe19c55cd72e7b9469e9fccc94
Closes-bug: #1732462
(cherry picked from commit 552d6fd33a4a2534eea6ca71765ab9936f5920b7)