SMBFS: remove deprecated config options
Some driver specific config options have been marked as deprecated
in Pike and are now being removed. The equivalent generic options
are available.

Note that the 'smbfs_sparsed_volumes' config option was only used
by the Linux SMB driver, which was removed in Pike. By mistake,
it was copied over to the Windows SMB driver, which was inheriting
it.

Change-Id: I23def8b2095dcf3945d2e0e1b4e5efe178cef9be