handle no default_cgsnapshot_type
If default_cgsnapshot_type has been deleted, group operations for
drivers relying on the generic volume group implementation were failing
with a KeyError while attempting to determine if the targeted group was
default_cgsnapshot_type. This fixes that by using a comparison that
allows for the possibility that default_cgsnapshot_type does not
exist.

Change-Id: Ic220ea4d8a6e83293e5237acfb07b1a417293803
Closes-Bug: #1710937