Update api-ref to include volume_image_metadata
The detailed information for a volume can include "volume_image_metadata"
in the case where the volume was created from an image, or in the case
where the volume was created from a snapshot of such a volume.

Update the API ref to include the field so we accurately represent
the code.

Change-Id: If5fcdefaa10bfb00e60aa4559d3282d3c6a53752
Closes-Bug: #1732763