VMAX driver - Fix error handling and checks for generic volume groups
There are several issues with the implementation of generic volume
groups in the VMAX driver. There is no rollback implemented in case
of errors for create group from source or for deleting groups. There
can be issues if source groups have been modified since a snapshot
was taken. The provider_location is also not being updated in the
individual volume snapshot objects of a group snapshot. This patch
rectifies these issues, and also does some general refactoring of
the generic volume group code.

Change-Id: I3e6ac90b145015ccd820af3449b20e377c9f4651
Closes-bug: 1732488