Move legacy-cinder-tox-* jobs in-tree
The compliance and bandit-baseline jobs just call tox to run tests.
These are an easy step to move in-tree.

Since it is also easier to have conditions on when things run, this
also adds filtering so the compliance job will now only run when there
are changes in code that impact the drivers.

Change-Id: Idb9a1f39481f3ebb704b28abc062ae24414d7224
Needed-by: I3af7771ab3e8d5a192f466dcec36a2b41db08ac3
Needed-by: I4af1813c69a5ddccbde8b99ed15023aa407995b5