Fix policy documentation for os-show_image_metadata endpoint
The policy documentation for os-show_image_metadata endpoint
is inconsistent with the implementation.

In policy documentation, the endpoint is listed as enforcing
"volume_extension:volume_image_metadata": [0]

But in the code it is clear that no policy enforcement is performed: [1]

This PS removes the os-show_image_metadata entry from the policy
documentation for consistency.

[0] https://github.com/openstack/cinder/blob/9fbf236dc17d67404e2ce9ced517182716d72e6f/cinder/policies/volume_metadata.py#L102
[1] https://github.com/openstack/cinder/blob/0cf910d4345c000e8c306b1cb2b2dd291975cf71/cinder/api/contrib/volume_image_metadata.py#L123

Change-Id: Icfde885384d174dbf9b211faca3f31be0b41f232
Closes-Bug: #1733632