Update unity tests to use our test base
Deprecation warnings were being emitted due to use of assertRaisesRegex,
even though a py2/3 wrapper was created in
I3cf8e29184a68a07d35fb121c816e40b89df393e. This was due to the Unity
tests inheriting from unittest.TestCase instead of our own base TestCase.

Switched test cases over to use cinder.test.TestCase gets rid of the
deprecation warning for assertRaisesRegex.

Our base test installs a logging fixture to make sure log messages are
OK. This uncovered that the driver code was trying to pass an object in
as if it were a dict of formatting values, resulting in an error. This
patch fixes this my creating a dict of the needed values to pass in so
the tests will pass under the new inheritance.

Change-Id: I2a6643ab4827dae244212c90c73402ff6d61cf76