Check for migrated UUID in SolidFire delete
Migration of volumes kinda screws up the use of ID as the
reference to use for deleting volumes.

If a volume is migrated, the UUID (ID) gets swapped and no
longer matches the UUID the volume had when it was initially
created on the back end device.  The result is that deletion
of that volume fails on the back end because the handle to the
new UUID doesn't exist.

On the other hand, you also have to be careful in dealing with
and cleaning up volumes migrated off of your back end device for
the same types of reasons.

This change just adds an additional check in the SF driver so that
we are querying the back end device for the ID and the name_id.

This way we make sure if the volume is on our device we clean it up.

Change-Id: I60f8e8864f236fb92dd03560576377f1a0bdd82b
Closes-Bug: #1733967
(cherry picked from commit 0e6ee1b1df38059f338ebd93c605dfb51a93a8a5)