VMAX driver - Errors extending replicated volumes
When attempting to extend multiple volumes concurrently, the operation
can fail due to insufficient locking on the RDF group. Additionally,
attempting to delete a volume after an earlier failure can fail as the
code does not check if the volume is already suspended. We also cannot
support extending SRDF/Metro volumes. This patch rectifies these issues.

Change-Id: Ib30f2a8f0a710f01ecb55ed753ecc41dbcdaad18
Closes-Bug: 1735456