Discrepancy in api-ref and code for create group type API
'name' is mentioned as an optional parameter in api-ref doc for
POST /v3/{project_id}/group_types API but in the code if we don't
pass 'name' it returns BadRequest. Also in create and update API,
'is_public' parameter is not added in api-ref document.

This patch fixes this discrepancy by updating the api-ref according
to the code.

Change-Id: I34e53ac72cbda9d8264e48e50f5a84ca09b850bd