Add shared_targets and service_uuid to volumes
This adds the last bits to the shared_targets and service_uuid
additions to volumes.  The object and db models are already
updated and this change simply introduces API MV 3.48 to
expose the new fields to the Volume view.

These fields are intended to be used for identifying if the
back end device serving a volume is using shared targets, and
provides the service_uuid to be used as a suggested/unique
handle for locking.

Change-Id: Ibd18c5f6c772f7265175a63ed2b598fcc3c87bc1