Fix discrepancy in api-ref for volume_types APIs
In the api ref document, for create volume_types 'extra_specs'
parameter is missing. And for parameter 'is_public', in code it
is used as 'os-volume-type-access:is_public'. For update volume_types
api as per api ref, 'name' and 'extra_specs' parameters are mandatory.
But in code 'name' is optional and 'extra_specs' implementation is not
done.

This patch fixes this discrepancy by updating the api-ref according to
the code.

Change-Id: I7faa711fc72cd81676812017a92ef40273f1bf7c