Deprecate cinder-manage logs commands
These commands search Cinder log files and
/var/log/messages to look for log messages related
to Cinder.

This is hacky and bound to be unreliable in
deployments which now may have log files configured
in different ways.

They also read entire log files into memory as
lists and reverse them, which is bound to be slow
on large deployments.

Deprecate these commands and leave log analysis
as a job for another tool.

Change-Id: I5672904383d41949b86cfdff41cfa0f0ae79b5ba