Deprecate CG APIs
This patch prints a deprecation message when CG APIs are used and
prompts users to swich to Generic Volume Group APIs instead. CG APIs
are also marked as deprecated in API reference docs. CG APIs will be
removed in a future release when it is appropriate and will be
decided by the Cinder team. This was communicated in Pike.

Change-Id: Ib6751fae6b5fb78de98a2ea62f507f9102f71b76