3PAR: Update CHAP on host record when volume is migrated to new compute host
When the nova instances are migrated from one compute host to another,
during the new VLUN creation step, the host record created on 3PAR has
the old chap credentials, but the volume metadata contains the newly
generated CHAP credentials. This fix will ensure the CHAP authentication
credentials on volume metadata and the host record on 3PAR are in sync
when the nova instances are live migrated.

Closes-Bug: #1737181

Change-Id: I2d4b7530b2d249ca193f7d0e87d529c9eb2dc25d