SMBFS: allow snapshot ro attach
This change will allow snapshots created by the SMBFS driver to be
attached in read-only mode. This will be used in order to backup
in-use volumes.

Note that writing to a snapshot would corrupt the differencing image
chain.

We'll need to locate snapshot backing files. The issue is that
in some cases, in-place locks may prevent us from querying the top
image from the chain. For this reason, we're going to store the
backing file information in the DB using the snapshot object
metadata field, at the same time preserving backwards compatibility.

Fake volume/snapshot object usage throughout the smbfs driver unit
tests had to be updated due to the object changes performed by it.

Partial-Implements: bleuprint windows-smb-backup

Change-Id: Ideaacbf9d160f400bef53825103b671127252789