SMBFS: fix creating volume from snapshot
In some cases, attempting to create a volume from an in-use snapshot
fails because of Hyper-V locks.

The reason is that in order to locate the backing image that needs
to be *cloned*, we may need to query the currently attached image,
which is not allowed.

In order to avoid this, we're going to store the backing file
information in the DB using the snapshot object metadata field,
at the same time preserving backwards compatibility.

Note that we cannot rely on model changes passed to the manager as
snapshots are used internally in a few cases in which the manager
would not expect snapshot object changes. Also, when deleting
snapshots, we're updating the metadata for more recent snapshots
as the backing file changes.

Fake volume/snapshot object usage throughout the smbfs driver unit
tests had to be updated due to the object changes performed by it.

Change-Id: I88e9dad5cf36621c4b86acc7b7e972850e061119
Closes-Bug: #1694635