Store host connector in volume_attachment.connector column
The attachment_specs table's key and value columns are strict
strings, which means things like a wwpns list value for a fibrechannel
connector can't get stored there and results in a DB error during
attach with the new volume attach flow in Nova.

The attachment_specs table is arguably not the best way to store
this data, which is just a dict like the connection_info.

A better way to store this is as a serialized json blob on the
volume_attachment record itself.

This patch adds the database migration to add the column and
an online data migration routine to migrate existing attachment_specs
entries when a volume attachment object is loaded from the database.

The volume manager attachment_update flow is changed to store
new connector attachments in the volume_attachment table directly.

An online data migration hook for the CLI will be added in a follow
up change.

Change-Id: Ica1f0e06adf0afcf740aad8cdc8d133ada1760c8
Closes-Bug: #1737724