Don't call driver.terminate_connection if there is no connector
It's possible to attach a volume to a shelved offloaded instance,
which is not on a compute host - which means no host connector,
and then detach that volume while the instance is unshelved. There
is a test in Tempest just for this scenario.

When nova calls attachment_delete on detach, the cinder volume
manager is calling the driver.terminate_connection method regardless
of there being a host connector, and some drivers blow up on that.

This used to "work" with the pre-3.27 attach flow in nova because
in the detach case where there was no connector, nova would simply
not call os-terminate_connection, but with the new 3.27+ flow for
attach, nova only calls attachment_delete regardless of there
being a connector.

This change simply checks to see if the attachment has a connector
and if not, it avoids the call to the driver to terminate a connection
which does not exist.

Change-Id: I496e45608798a6a5d9606f9594feeb8b60855d1a
Closes-Bug: #1738254